acq400_stream now collects recv data up to a user specified file size
Data is collected by sock.recv and is stored until it reaches a user specified size, where it is then written to a dat file. For future commits it would be useful to have a time feature where the data is captured until EITHER time condition is met or file size exceeds user specification.